requirements.txt
dacite==1.6.0
PyYAML==6.0.1
SQLAlchemy==1.4.20
psycopg2-binary<=3.0.0; sys_platform == 'darwin'
psycopg2<=3.0.0; sys_platform == 'linux'
pytz==2024.1
click==8.1.3
pydantic==1.10.15
python-telegram-bot>=13.3
grpcio==1.63.0
protobuf~=5.28.2
grpc-interceptor==0.15.0
protovalidate==0.5.0
